requirements.txt
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
flask-restx==1.1.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.2
six==1.16.0
werkzeug==2.3.7
zipp==3.16.2
